+++ db/upgrade.php
@@ -17,24 +17,29 @@
 // phpcs:disable moodle.Files.BoilerplateComment.CommentEndedTooSoon
 
 /**
- * Custom home pages plugin version.
+ * Plugin tool_muhome upgrades.
  *
- * @package     tool_muhome
- * @copyright   2025 Petr Skoda
- * @license     https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ * @package    tool_muhome
+ * @copyright  2026 Petr Skoda
+ * @license    https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-defined('MOODLE_INTERNAL') || die();
+/**
+ * Upgrade muhome.
+ *
+ * @param mixed $oldversion
+ * @return true
+ */
+function xmldb_tool_muhome_upgrade($oldversion): bool {
+    global $DB;
 
-/** @var stdClass $plugin */
-$plugin->component = 'tool_muhome';
-$plugin->version = 2026013145;
-$plugin->requires = 2024100700;
-$plugin->maturity = MATURITY_BETA;
-$plugin->supported = [405, 405];
-$plugin->incompatible = 500;
-$plugin->release = 'mu-4.5.8-04+';
+    $dbman = $DB->get_manager();
 
-$plugin->dependencies = [
-    'tool_mulib' => 2026012545,
-];
+    if ($oldversion < 2026013145) {
+        $table = new xmldb_table('tool_muhome_cohortvisible');
+        $dbman->rename_table($table, 'tool_muhome_page_cohortvisible');
+        upgrade_plugin_savepoint(true, 2026013145, 'tool', 'muhome');
+    }
+
+    return true;
+}